Fix utils race and error while in `-race` mode
- Using a safe map instead of a simple map for `TestInParallel` and
  `TestInParallelError` to be sure we don't have a data race while
  running the tests.
- Putting these two (`TestInParallel` and `TestInParallelError`) into a
  separate test file with build tag `!race`. `InParallel` relies on
  `sync.Pool` and `sync.Pool` is no-op in `-race` mode ; this means
  `TestInParallel` is always OK, `TestInParallelError` is always KO when
  race detector is on.
